Component Structure Accessing State:
useSelector: This hook is used to access the Redux store's state. It selects the posts array from the post slice of the store.
const {posts} = useSelector(store => store.post);: Destructures the posts array from the post slice of the Redux state. This array contains the posts that need to be rendered.
Rendering the List:

<div>: The container element that wraps the list of posts.
posts.map((post) => <Post key={post._id} post={post} />): Iterates over the posts array and renders a Post component for each post object.
post._id: Used as a unique key for each Post component, which helps React efficiently update and manage the list.
post={post}: Passes the individual post object as a prop to the Post component.