Component State and Hooks State Variables:
text: Tracks the content of the comment input field.
open: Determines whether the comment dialog is open.
liked: Boolean indicating if the post is liked by the current user.
postLike: The number of likes on the post.
comment: Array of comments associated with the post.
Redux Hooks:

useSelector: Used to access the Redux store's state, specifically the user and posts data.
useDispatch: Used to dispatch actions to the Redux store.