Event Handlers changeEventHandler: Updates the text state when the input field value changes. It trims whitespace and updates the state accordingly.
likeOrDislikeHandler: Toggles the like status of the post:

Sends a request to the server to either like or dislike the post based on the current state.
Updates the local state (postLike and liked) and Redux store (posts slice) with the new like status.
Displays a success message using toast.
commentHandler: Adds a new comment to the post:

Sends a request to the server with the comment text.
Updates the local state (comment) and Redux store (posts slice) with the new comment.
Displays a success message using toast.
deletePostHandler: Deletes the post:

Sends a request to the server to delete the post.
Updates the local state (posts slice) to remove the deleted post.
Displays a success message using toast.
bookmarkHandler: Toggles bookmarking of the post:

Sends a request to the server to bookmark the post.
Displays a success message using toast.