Handlers sendMessageHandler:
Parameters: receiverId (ID of the user receiving the message).
Functionality: Sends a POST request to the server to send a message to the selected user. On success, it updates the Redux store with the new message and clears the input field.
useEffect:

Purpose: Cleans up by dispatching setSelectedUser(null) when the component is unmounted, ensuring that no user is selected when the chat page is closed.